Eye Spy a fix!
made eye spy now work with Better End strongholds and with End Remastered stronghold replacement structure if they are on. To do this, I made a custom criteria and predicate that is structure safe. if the structure passed in doesnt exist in the structure registry, it makes the predicate return false instead of instantly granting the advancement. The feature field for the custom predicate will use the structure registry instead of the bimap to resolve what structure to find